+++ purchases/src/test/java/com/revenuecat/purchases/PurchasesCommonTest.kt
@@ -1386,194 +1386,6 @@ internal class PurchasesCommonTest: BasePurchasesTest() {
 
     // endregion
 
-    // region canMakePayments
-
-    @Test
-    fun `when calling canMakePayments and billing service disconnects, return false`() {
-        var receivedCanMakePayments = true
-        val mockLocalBillingClient = mockk<BillingClient>(relaxed = true)
-
-        val listener = setUpMockBillingClientBuilderAndListener(mockLocalBillingClient)
-
-        Purchases.canMakePayments(mockContext, listOf()) {
-            receivedCanMakePayments = it
-        }
-        listener.captured.onBillingServiceDisconnected()
-        assertThat(receivedCanMakePayments).isFalse
-        verify(exactly = 1) { mockLocalBillingClient.endConnection() }
-    }
-
-    @Test
-    fun `canMakePayments with no features and OK BillingResponse returns true`() {
-        var receivedCanMakePayments = false
-        val mockLocalBillingClient = mockk<BillingClient>(relaxed = true)
-        val listener = setUpMockBillingClientBuilderAndListener(mockLocalBillingClient)
-
-        Purchases.canMakePayments(mockContext, listOf()) {
-            receivedCanMakePayments = it
-        }
-
-        listener.captured.onBillingSetupFinished(BillingClient.BillingResponseCode.OK.buildResult())
-
-        assertThat(receivedCanMakePayments).isTrue
-    }
-
-    @Test
-    fun `when no play services, canMakePayments returns false`() {
-        var receivedCanMakePayments = true
-        val mockLocalBillingClient = mockk<BillingClient>(relaxed = true)
-        val listener = setUpMockBillingClientBuilderAndListener(mockLocalBillingClient)
-
-        every { mockLocalBillingClient.endConnection() } throws mockk<IllegalArgumentException>()
-        Purchases.canMakePayments(mockContext, listOf(BillingFeature.SUBSCRIPTIONS)) {
-            receivedCanMakePayments = it
-        }
-        listener.captured.onBillingSetupFinished(BillingClient.BillingResponseCode.FEATURE_NOT_SUPPORTED.buildResult())
-        assertThat(receivedCanMakePayments).isFalse
-        verify(exactly = 1) { mockLocalBillingClient.endConnection() }
-    }
-
-    @Test
-    fun `when calling canMakePayments, enablePendingPurchases is called`() {
-        val mockLocalBillingClient = mockk<BillingClient>(relaxed = true)
-        every {
-            mockLocalBillingClient.isFeatureSupported(BillingClient.FeatureType.SUBSCRIPTIONS)
-        } returns BillingClient.BillingResponseCode.FEATURE_NOT_SUPPORTED.buildResult()
-
-        mockkStatic(BillingClient::class)
-        val mockBuilder = mockk<BillingClient.Builder>(relaxed = true)
-        every { BillingClient.newBuilder(any()) } returns mockBuilder
-        every { mockBuilder.setListener(any()) } returns mockBuilder
-        every { mockBuilder.enablePendingPurchases() } returns mockBuilder
-        every { mockBuilder.build() } returns mockLocalBillingClient
-        every { mockLocalBillingClient.startConnection(any()) } just Runs
-
-        Purchases.canMakePayments(mockContext, listOf(BillingFeature.SUBSCRIPTIONS)) {}
-        verify(exactly = 1) { mockBuilder.enablePendingPurchases() }
-    }
-
-    fun `canMakePayments returns true for Amazon configurations`() {
-        purchases.purchasesOrchestrator.appConfig = AppConfig(
-            mockContext,
-            false,
-            false,
-            PlatformInfo("", null),
-            null,
-            Store.AMAZON
-        )
-        Purchases.canMakePayments(mockContext, listOf()) {
-            assertThat(it).isTrue()
-        }
-    }
-
-    @Test
-    fun `when billing is not supported, canMakePayments is false`() {
-        var receivedCanMakePayments = true
-        val mockLocalBillingClient = mockk<BillingClient>(relaxed = true)
-        val listener = setUpMockBillingClientBuilderAndListener(mockLocalBillingClient)
-
-        every {
-            mockLocalBillingClient.isFeatureSupported(BillingClient.FeatureType.SUBSCRIPTIONS)
-        } returns BillingClient.BillingResponseCode.OK.buildResult()
-
-        mockHandlerPost()
-
-        Purchases.canMakePayments(mockContext, listOf(BillingFeature.SUBSCRIPTIONS)) {
-            receivedCanMakePayments = it
-        }
-
-        listener.captured.onBillingSetupFinished(BillingClient.BillingResponseCode.BILLING_UNAVAILABLE.buildResult())
-
-        assertThat(receivedCanMakePayments).isFalse
-        verify(exactly = 1) { mockLocalBillingClient.endConnection() }
-    }
-
-    @Test
-    fun `when feature is not supported, canMakePayments is false`() {
-        var receivedCanMakePayments = true
-        val mockLocalBillingClient = mockk<BillingClient>(relaxed = true)
-        val listener = setUpMockBillingClientBuilderAndListener(mockLocalBillingClient)
-
-        every {
-            mockLocalBillingClient.isFeatureSupported(BillingClient.FeatureType.SUBSCRIPTIONS)
-        } returns BillingClient.BillingResponseCode.FEATURE_NOT_SUPPORTED.buildResult()
-
-        Purchases.canMakePayments(mockContext, listOf(BillingFeature.SUBSCRIPTIONS)) {
-            receivedCanMakePayments = it
-        }
-
-        listener.captured.onBillingSetupFinished(BillingClient.BillingResponseCode.OK.buildResult())
-
-        assertThat(receivedCanMakePayments).isFalse
-        verify(exactly = 1) { mockLocalBillingClient.endConnection() }
-    }
-
-    @Test
-    fun `when one feature in list is not supported, canMakePayments is false`() {
-        var receivedCanMakePayments = true
-        val mockLocalBillingClient = mockk<BillingClient>(relaxed = true)
-        val listener = setUpMockBillingClientBuilderAndListener(mockLocalBillingClient)
-
-        every {
-            mockLocalBillingClient.isFeatureSupported(BillingClient.FeatureType.SUBSCRIPTIONS)
-        } returns BillingClient.BillingResponseCode.FEATURE_NOT_SUPPORTED.buildResult()
-
-        every {
-            mockLocalBillingClient.isFeatureSupported(BillingClient.FeatureType.SUBSCRIPTIONS_UPDATE)
-        } returns BillingClient.BillingResponseCode.OK.buildResult()
-
-        Purchases.canMakePayments(
-            mockContext,
-            listOf(
-                BillingFeature.SUBSCRIPTIONS,
-                BillingFeature.SUBSCRIPTIONS_UPDATE
-            )
-        ) {
-            receivedCanMakePayments = it
-        }
-
-        listener.captured.onBillingSetupFinished(BillingClient.BillingResponseCode.OK.buildResult())
-
-        assertThat(receivedCanMakePayments).isFalse
-        verify(exactly = 1) { mockLocalBillingClient.endConnection() }
-    }
-
-    @Test
-    fun `when single feature is supported and billing is supported, canMakePayments is true`() {
-        var receivedCanMakePayments = false
-        val mockLocalBillingClient = mockk<BillingClient>(relaxed = true)
-        val listener = setUpMockBillingClientBuilderAndListener(mockLocalBillingClient)
-
-        every {
-            mockLocalBillingClient.isFeatureSupported(BillingClient.FeatureType.SUBSCRIPTIONS)
-        } returns BillingClient.BillingResponseCode.OK.buildResult()
-
-        mockHandlerPost()
-        Purchases.canMakePayments(mockContext, listOf(BillingFeature.SUBSCRIPTIONS)) {
-            receivedCanMakePayments = it
-        }
-
-        listener.captured.onBillingSetupFinished(BillingClient.BillingResponseCode.OK.buildResult())
-
-        assertThat(receivedCanMakePayments).isTrue
-        verify(exactly = 1) { mockLocalBillingClient.endConnection() }
-    }
-
-    @Test
-    fun `when feature list is empty, canMakePayments does not check billing client for feature support`() {
-        val mockLocalBillingClient = mockk<BillingClient>(relaxed = true)
-        val listener = setUpMockBillingClientBuilderAndListener(mockLocalBillingClient)
-
-        Purchases.canMakePayments(mockContext, listOf()) {}
-
-        listener.captured.onBillingSetupFinished(BillingClient.BillingResponseCode.OK.buildResult())
-
-        verify(exactly = 0) { mockLocalBillingClient.isFeatureSupported(any()) }
-        verify(exactly = 1) { mockLocalBillingClient.endConnection() }
-    }
-
-    // endregion
-
     // region teardown
 
     @Test
